test(runtime): add headless serve parity, extract paired-client harness, register view-isolation gate (STA-2802)
